fix(web): linter & build
