--- web/src/pages/Register/signup.scss
@@ -1,203 +1,0 @@
-
-.signup-container{
-    display: flex;
-    width: 100%;
-    height: 100vh;
-}
-.welcom{
-  font-family: 'Roboto';
-  font-size: 20px;
-}
-.signup-text{
-  background-color: #F4F5F8;
-  display: flex;
-  width: 30%;
-  margin: 0;
-  padding: 2rem;
-  justify-content: center;
-  text-align: left;
-  align-items: center;
-}
-.text-box{
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  padding: 40px 40px;
-}
-
-.alpha{ 
-  font-family: 'Poppins-Black';
-  font-style: normal;
-  font-weight: 900;
-  font-size: 28px;
-  line-height: 32px;
-  display: flex;
-  align-items: center;
-
-  color: rgba(0, 0, 0, 0.87);
-  flex: none;
-  order: 0;
-  flex-grow: 0;
-}
-
-.betta{ 
-  font-family: 'Roboto-Light';
-  font-size: 26px;
-  font-weight: 400;
-  line-height: 56px;
-  letter-spacing: 0px;
-  text-align: left;
-  display:flex;
-  align-items: center;
-  letter-spacing: 0.25px;
-
-  color: rgba(0, 0, 0, 0.87);
-  
-  flex: none;
-  order: 0;
-  flex-grow: 0;
-}
-
-.charlie{ 
-  font-family: 'Roboto-Light';
-  font-style: normal;
-  font-size: 16px;
-}
-
-.signup-form{
-    display: flex;
-    width: 70%;
-    background-color: white;
-    justify-content: center;
-    align-items: center;
-}
-
-    
-  .wrapper-log {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  .form-wrapper-log {
-    display: flex;
-    flex-direction: column;
-    width: 60%;
-    max-width: 80%;
-    min-width: 100px;
-    min-height: 400px;
-    padding: 20px 40px;
-  }
-  
- 
-  form {
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-  }  
-  
-  label {
-    margin-bottom: 0.5em;
-    color: black;
-    font-weight: lighter;
-  }
-
-  .button {
-    min-width: 100%;
-    cursor: pointer;
-    margin-right: 0.25em;
-    margin-top: 0.5em;
-    padding:  0.938em;
-    border: none;
-    border-radius: 4px;
-    background-color:  #7C44F3;
-    color: #fefefe;
-  }
-  .button:hover {
-    background-color: rgb(6, 73, 11);
-    color: #fefefe;
-  }
-   
-  .Confirmpassword,
-  .email,
-  .password {
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 15px;
-    width: 100%;
-  }
-   
-  .submit {
-    margin-left: 20px;
-    width: 94%;
-    display: flex;
-    flex-wrap: wrap;
-  }
-
-  .input-block {
-    position: relative;
-    width: 100%;
-    margin: 20px;
-  }
-
-  .input-block input {
-    padding: 10px 10px;
-    border-radius: 5px;
-
-    width: 100%;
-    height: 50px;
-    
-    font-size: 14px;
-    font-family: "Poppins-Regular";
-    border-radius: 5px;
-    outline: none;
-    border: 1px solid #9A9A9A;
-  }
-
-  .placeholder {
-    font-family: Roboto;
-    font-size: 13px;
-    font-weight: 400;
-    line-height: 24px;
-    letter-spacing: 0.15000000596046448px;
-    text-align: left;
-
-
-    position: absolute;
-    margin: 17px 0;
-    padding: 0 4px;
-    
-    color:  #9A9A9A;
-    display: flex;
-    align-items: center;
-    
-    top: 0;
-    left: 15px;
-    transition: all 0.2s;
-    transform-origin: 0% 0%;
-    background: none;
-    pointer-events: none;
-    transform: scale(0.8) translateY(-30px);
-    background: #fff;
-  }
-.log-box{
-  margin: 10px 170px;
-  font-family: 'Roboto-Light';
-  font-size: 13px;
-}
-.txt-color{
-  color: #7C44F3;
-}
-
-.i {
-  position: absolute;
-  top: 32%;
-  right: 2%;
-}
-.i:hover {
-  color: #7C44F3;
-  cursor: pointer;
-}
