+++ web/src/pages/Vulnerability/Vulnerability.scss
